pyproject.toml
[tool.poetry]
name = "kotano-gendiff"
version = "0.7.0"
description = "Find difference between two files"
authors = ["kotano"]

packages = [
    { include = "gendiff" },
]

[tool.poetry.scripts]
gendiff = "gendiff.scripts.run_gendiff:main"

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-cov = "^2.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
